backend/requirements.txt
Flask
flask-cors
python-dotenv
pyrebase4
werkzeug
gunicorn